Restructure tests, remove old doc comments.
