+++ Tests/RequestTests.swift
@@ -734,24 +734,6 @@ final class RequestResponseTestCase: BaseTestCase {
         XCTAssertNotNil(response2?.value)
         XCTAssertNotNil(response3?.value)
     }
-
-    func testThatRequestsCanHaveCustomTimeoutValueSet() {
-        // Given
-        let url = URL.makeHTTPBinURL(path: "delay/1")
-        let expect = expectation(description: "request completed")
-        var response: AFDataResponse<Data?>?
-
-        // When
-        AF.request(url) {
-            $0.timeoutInterval = 0.01
-        }
-        .response { response = $0; expect.fulfill() }
-
-        waitForExpectations(timeout: timeout)
-
-        // Then
-        XCTAssertEqual((response?.error?.underlyingError as? URLError)?.code, .timedOut)
-    }
 }
 
 // MARK: -
